Renaming the package from com.arkbuilders to dev.arkbuilders
