--- app/src/main/java/com/arkbuilders/arkdrop/presentation/navigation/NavRegister.kt
@@ -1,28 +1,0 @@
-package com.arkbuilders.arkdrop.presentation.navigation
-
-import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
-import com.arkbuilders.arkdrop.Greeting
-import com.arkbuilders.arkdrop.presentation.feature.filestransfers.FilesTransferScreen
-import com.arkbuilders.arkdrop.presentation.feature.settings.SettingsScreen
-import com.arkbuilders.arkdrop.presentation.feature.transferconfirmation.TransferConfirmationScreen
-import com.arkbuilders.arkdrop.presentation.feature.transferprogress.TransferProgressScreen
-
-fun NavGraphBuilder.navRegistration(navController: NavController) {
-    composable(TransfersDestination.route) {
-        FilesTransferScreen(navController = navController)
-    }
-    composable(HistoryDestination.route) {
-        Greeting(name = "HistoryDestination")
-    }
-    composable(SettingsDestination.route) {
-        SettingsScreen()
-    }
-    composable(TransferConfirmationDestination.route) {
-        TransferConfirmationScreen(navController = navController)
-    }
-    composable(TransferProgressDestination.route) {
-        TransferProgressScreen(navController = navController)
-    }
-}
