--- app/src/main/java/com/arkbuilders/arkdrop/presentation/permission/PermissionManager.kt
@@ -1,76 +1,0 @@
-package com.arkbuilders.arkdrop.presentation.permission
-
-import android.Manifest
-import android.content.Context
-import android.content.pm.PackageManager
-import android.os.Build
-import androidx.activity.result.ActivityResultLauncher
-import androidx.annotation.RequiresApi
-import androidx.core.content.ContextCompat
-
-object PermissionManager {
-
-    private lateinit var requestPermissionLauncher: ActivityResultLauncher<Array<String>>
-
-    fun initialize(requestPermissions: ActivityResultLauncher<Array<String>>) {
-        requestPermissionLauncher = requestPermissions
-    }
-
-    @RequiresApi(Build.VERSION_CODES.UPSIDE_DOWN_CAKE)
-    private val requiredPermissionApi34 = arrayOf(
-        Manifest.permission.READ_MEDIA_IMAGES,
-        Manifest.permission.READ_MEDIA_VIDEO,
-        Manifest.permission.READ_MEDIA_VISUAL_USER_SELECTED,
-        Manifest.permission.CAMERA,
-    )
-
-    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
-    private val requiredPermissionApi33 = arrayOf(
-        Manifest.permission.READ_MEDIA_IMAGES,
-        Manifest.permission.READ_MEDIA_VIDEO,
-        Manifest.permission.CAMERA,
-    )
-
-    fun requestPermission(context: Context) {
-        if (allPermissionGranted(context)) {
-            return
-        }
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-            requestPermissionLauncher.launch(requiredPermissionApi34)
-        } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            requestPermissionLauncher.launch(requiredPermissionApi33)
-        }
-    }
-
-    fun isCameraPermissionGranted(context: Context) = run {
-        ContextCompat.checkSelfPermission(
-            context, Manifest.permission.CAMERA
-        ) == PackageManager.PERMISSION_GRANTED
-    }
-
-    fun requestCameraPermission() {
-        requestPermissionLauncher.launch(
-            arrayOf(
-                Manifest.permission.CAMERA,
-            )
-        )
-    }
-
-    private fun allPermissionGranted(context: Context): Boolean {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-            return requiredPermissionApi34.all {
-                ContextCompat.checkSelfPermission(
-                    context, it
-                ) == PackageManager.PERMISSION_GRANTED
-            }
-        } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            return requiredPermissionApi33.all {
-                ContextCompat.checkSelfPermission(
-                    context, it
-                ) == PackageManager.PERMISSION_GRANTED
-            }
-        }
-        return false
-    }
-
-}
